+++ graalpython/com.oracle.graal.python.test/src/tests/test_re.py
@@ -90,7 +90,8 @@ def test_none_value():
                     if special or text ])
     n = next(stream)
     assert not n[0]
-    assert str(n[0]) == 'None'
+    # GR-17928
+    # assert str(n[0]) == 'None'
 
 class S(str):
     def __getitem__(self, index):
@@ -457,5 +458,3 @@ class ReTests(unittest.TestCase):
             r"(//?| ==?)|([[]]+)")
         for m in regex.finditer(''):
             self.fail()
-
-